requirements.txt
Flask==3.1.2
Quart==0.19.9
SQLAlchemy==2.0.31
aiosqlite==0.20.0
python-dotenv==1.0.1
argon2-cffi==23.1.0
PyJWT==2.8.0
itsdangerous==2.2.0
quart-cors==0.7.0
click==8.1.7
